webapp/backend/requirements.txt
flask
flask-cors
redis
python-dotenv
gunicorn